Move Twig related classes to Templating\Twig namespace
