+++ Templating/Twig/TwigEnvironmentTrait.php
@@ -9,8 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Lolautruche\EzCoreExtraBundle\Templating;
+namespace Lolautruche\EzCoreExtraBundle\Templating\Twig;
 
+use Lolautruche\EzCoreExtraBundle\Templating\TemplateNameResolverInterface;
 use Lolautruche\EzCoreExtraBundle\Templating\Twig\DebugTemplate;
 
 trait TwigEnvironmentTrait
